refactor(app): split useD3Lollipop into d3-lollipop module directory (#346)
