+++ app/src/composables/d3-lollipop/index.ts
@@ -1,0 +1,9 @@
+/**
+ * d3-lollipop/index.ts
+ *
+ * Public barrel for the d3-lollipop module. Only the composable and its
+ * option/state types are part of the public composables API.
+ */
+
+export { useD3Lollipop } from './useD3Lollipop';
+export type { LollipopOptions, D3LollipopState, PlotMargin } from './useD3Lollipop';
